o/auth, seclog: avoid seclog type SnapdUserState that duplicates auth type UserState
